+++ src/InstancesDataProvider.ts
@@ -3,6 +3,8 @@ import axios from "axios";
 import { ADD_INSTANCE } from './consts/Commands';
 import { mapMergeRequestToTreeItem, mapInstanceToTreeItem } from './utils/mappers';
 import { InstanceTreeItem } from './impl/TreeItem/InstanceTreeItem';
+import { isUsingApprovals } from './utils/isUsingApprovals';
+import { GitlabMergeRequest } from './interfaces/GitLabMergeRequest';
 
 export class InstancesDataProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
   private _onDidChangeTreeData = new vscode.EventEmitter<InstanceTreeItem | undefined>();
@@ -21,21 +23,35 @@ export class InstancesDataProvider implements vscode.TreeDataProvider<vscode.Tre
     return element;
   }
 
+  private getAxiosConfig(host: string) {
+    return {
+      headers: {
+        'Private-Token': this.context.globalState.get(host),
+      }
+    };
+  }
+
+  async getApprovals(host: string, projectId: number, mergeRequestIid: number): Promise<number> {
+    const endpoint = `https://${host}/api/v4/projects/${projectId}/merge_requests/${mergeRequestIid}/approvals`
+
+    const { data: { approvals_required, approvals_left } } = await axios.get(endpoint, this.getAxiosConfig(host));
+
+    return approvals_required - approvals_left;
+  }
+
   async getChildren(element?: InstanceTreeItem): Promise<vscode.TreeItem[]> {
     // If there is an element, we are inside and instance, so we fetch the
     // opened Merge Request to get the different data.
     if (element) {
-      const response = await axios.get(
-        `https://${element.label}/api/v4/merge_requests?scope=created_by_me&state=opened`,
-        {
-          headers: {
-            'Private-Token': this.context.globalState.get(element.label),
-          }
-        }
+      const host = element.label;
+      const endpoint = `https://${host}/api/v4/merge_requests?scope=created_by_me&state=opened`;
+
+      const { data: mergeRequests } = await axios.get(
+        endpoint, this.getAxiosConfig(host),
       );
 
       // If no Merge Request is available, show a TreeItem with an explanation.
-      if (response.data.length === 0) {
+      if (mergeRequests.length === 0) {
         return [
           new vscode.TreeItem(
             'No merge request available for this instance',
@@ -43,7 +59,19 @@ export class InstancesDataProvider implements vscode.TreeDataProvider<vscode.Tre
         ];
       }
 
-      return response.data.map(mapMergeRequestToTreeItem);
+      if (isUsingApprovals()) {
+        const mergeRequestWithApprovals: Array<Promise<GitlabMergeRequest>> = mergeRequests.map(async (mergeRequest: GitlabMergeRequest) => {
+          const approvals = await this.getApprovals(host, mergeRequest.project_id, mergeRequest.iid);
+          mergeRequest.upvotes = approvals;
+          return mergeRequest;
+        });
+
+        const mergeRequestValues = await Promise.all(mergeRequestWithApprovals);
+
+        return mergeRequestValues.map(mapMergeRequestToTreeItem);
+      }
+
+      return mergeRequests.map(mapMergeRequestToTreeItem);
     } else {
       const instances = this.context.globalState
         .get(this.instancesKey, [])