+++ src/interfaces/GitLabMergeRequest.ts
@@ -1,4 +1,6 @@
 export interface GitlabMergeRequest {
+  iid: number;
+  project_id: number;
   title: string;
   upvotes: number;
   web_url: string;